+++ module/attribute/src/Application/Controller/Api/Attribute/AttributeValidationAction.php
@@ -52,7 +52,6 @@ class AttributeValidationAction
      * )
      *
      * @IsGranted("PRODUCT_ATTRIBUTE_POST_VALIDATION")
-     * @IsGranted("edit", subject="language")
      *
      * @SWG\Tag(name="Attribute")
      * @SWG\Parameter(