+++ module/core/src/Application/Model/LanguageTree/LanguageTreeUpdateFormModel.php
@@ -8,6 +8,7 @@ declare(strict_types=1);
 
 namespace Ergonode\Core\Application\Model\LanguageTree;
 
+use Ergonode\Core\Application\Validator as CoreAssert;
 use Symfony\Component\Validator\Constraints as Assert;
 
 class LanguageTreeUpdateFormModel
@@ -15,6 +16,8 @@ class LanguageTreeUpdateFormModel
     /**
      * @Assert\Valid()
      * @Assert\NotBlank(message="Languages is required")
+     *
+     * @CoreAssert\LanguageTreeLanguageRelation()
      */
     public ?LanguageTreeNodeFormModel $languages;