+++ module/core/src/Application/Validator/LanguageTreeLanguageRelation.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Copyright © Ergonode Sp. z o.o. All rights reserved.
+ * See LICENSE.txt for license details.
+ */
+
+declare(strict_types=1);
+
+namespace Ergonode\Core\Application\Validator;
+
+use Symfony\Component\Validator\Constraint;
+
+/**
+ * @Annotation
+ */
+class LanguageTreeLanguageRelation extends Constraint
+{
+    public string $message = 'Languages {{ languages }} must remain on the tree.';
+}